# Less norm computations in token similarity