I've removed the redundant vector norm computation in token similarity (was previously computed twice).

## Description

Before the change:

```
In [7]: doc = nlp('similar similarities')

In [8]: %timeit doc[0].similarity(doc[1])
38.8 µs ± 576 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

After the change:

```
In [3]: doc = nlp('similar similarities')

In [4]: %timeit doc[0].similarity(doc[1])
23.4 µs ± 131 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

### Types of change

Enhancement/speedup.

## Checklist
- [x] I have submitted the spaCy Contributor Agreement.
- [x] I ran the tests, and all new and existing tests passed.
- [x] My changes don't require a change to the documentation, or if they do, I've added all required information.