@honnibal 2018-09-05
Thanks! Not sure how I missed this. Surprised at the nice speed-up, too.